# Conflicts:
#	.vs/Fubon.Treasury/v14/.suo
#	Treasury.Web/Treasury.Web.csproj
#	Treasury.Web/bin/Treasury.Web.dll
#	Treasury.Web/bin/Treasury.Web.pdb
#	Treasury.Web/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	Treasury.Web/obj/Debug/RdlCompile.cache
#	Treasury.Web/obj/Debug/Treasury.Web.dll
#	Treasury.Web/obj/Debug/Treasury.Web.pdb

+++ Treasury.Web/Service/Actual/ItemImp.cs
@@ -12,6 +12,7 @@ using Treasury.Web.Enum;
 using System.ComponentModel;
 using System.Data.Entity.Infrastructure;
 using Treasury.WebControllers;
+using Treasury.Web.Controllers;
 
 /// <summary>
 /// 功能說明：金庫進出管理作業-金庫物品存取申請作業 重要物品
@@ -183,7 +184,8 @@ namespace Treasury.Web.Service.Actual
                             vItemId = x.ITEM_ID,
                             vlItem_Id = x.ITEM_ID,
                             vStatus = x.INVENTORY_STATUS,
-                            vPUT_Date = x.PUT_DATE?.ToString("yyyy/MM/dd"),
+                            vPUT_Date = x.PUT_DATE?.dateTimeToStr(),
+                            vGet_Date = x.GET_DATE?.dateTimeToStr(),
                             vAPLY_UID = x.APLY_UID,
                             vAPLY_UID_Name = emps.FirstOrDefault(y => y.USR_ID == x.APLY_UID)?.EMP_NAME?.Trim(),
                             vCHARGE_DEPT = x.CHARGE_DEPT,
@@ -205,6 +207,19 @@ namespace Treasury.Web.Service.Actual
                             vItemImp_MEMO_AFT = x.MEMO_AFT,
                             vLast_Update_Time = x.LAST_UPDATE_DT
                         }).ToList());
+
+                    if (searchModel.vTreasuryIO == "N") //取出
+                    {
+                        if (result.Any())
+                        {
+                            var itemIds = result.Select(x => x.vItemId).ToList();
+                            var uids = GetAplyUidNameByIMPO(itemIds);
+                            result.ForEach(x =>
+                            {
+                                x.vGet_Uid_Name = uids.FirstOrDefault(y => y.itemId == x.vItemId)?.getAplyUidName;
+                            });
+                        }
+                    }
                 }
                 else
                 {
@@ -218,7 +233,8 @@ namespace Treasury.Web.Service.Actual
                             vItemId = x.ITEM_ID,
                             vlItem_Id = x.ITEM_ID,
                             vStatus = x.INVENTORY_STATUS,
-                            vPUT_Date = x.PUT_DATE?.ToString("yyyy/MM/dd"),
+                            vPUT_Date = x.PUT_DATE?.dateTimeToStr(),
+                            vGet_Date = x.GET_DATE?.dateTimeToStr(),
                             vAPLY_UID = x.APLY_UID,
                             vAPLY_UID_Name = emps.FirstOrDefault(y => y.USR_ID == x.APLY_UID)?.EMP_NAME?.Trim(),
                             vCHARGE_DEPT = x.CHARGE_DEPT,
@@ -287,8 +303,17 @@ namespace Treasury.Web.Service.Actual
                             {
                                 #region 申請單紀錄檔
                                 _TAR = db.TREA_APLY_REC.First(x => x.APLY_NO == taData.vAplyNo);
-                                if (_TAR.APLY_STATUS != _APLY_STATUS) //申請紀錄檔狀態不是在表單申請狀態
-                                    _APLY_STATUS = Ref.AccessProjectFormStatus.A05.ToString(); //為重新申請案例
+                                if (CustodyAppr.Contains(_TAR.APLY_STATUS))
+                                {
+                                    _APLY_STATUS = CustodyConfirmStatus;
+                                    _TAR.CUSTODY_UID = AccountController.CurrentUserId; //保管單位直接帶使用者
+                                    _TAR.CUSTODY_DT = dt;
+                                }
+                                else
+                                {
+                                    if (_TAR.APLY_STATUS != _APLY_STATUS) //申請紀錄檔狀態不是在表單申請狀態
+                                        _APLY_STATUS = Ref.AccessProjectFormStatus.A05.ToString(); //為重新申請案例
+                                }
                                 _TAR.APLY_STATUS = _APLY_STATUS;
                                 _TAR.LAST_UPDATE_DT = dt;
 
@@ -373,7 +398,7 @@ namespace Treasury.Web.Service.Actual
                                             }
                                         }
                                     }
-                                    else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString())//取出
+                                    else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString() && (_APLY_STATUS != CustodyConfirmStatus))//取出
                                     {
                                         var _II = db.ITEM_IMPO.FirstOrDefault(x => x.ITEM_ID == item.vItemId);
                                         if (_II.LAST_UPDATE_DT > item.vLast_Update_Time)
@@ -441,7 +466,7 @@ namespace Treasury.Web.Service.Actual
                                     }));
                                     db.ITEM_IMPO.AddRange(inserts);
                                 }
-                                else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString())//取出
+                                else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString() && (_APLY_STATUS != CustodyConfirmStatus))//取出
                                 {
                                     db.OTHER_ITEM_APLY.RemoveRange(db.OTHER_ITEM_APLY.Where(x => x.APLY_NO == taData.vAplyNo).ToList());
                                     db.OTHER_ITEM_APLY.AddRange(updateItemIds.Select(x => new OTHER_ITEM_APLY()
